+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added enet library for moonlight-embedded-2.2.0
 - Solved a bug on xarcade where B and HOTKEY were sending the same event
 - Slide transition by default in ES
+- Power management switch support (power,reset and LED) for pin 3/5/6
 - Add ifconfig -a and /boot/recalbox-boot.conf in recalbox-support.sh
 
 ## [4.0.0-beta3] - 2016-04-19